Update reference to moved packages.
