+++ deeplearning/deeptune/opencl/heterogeneous_mapping/models/BUILD
@@ -83,7 +83,7 @@ py_library(
         ":ncc",
         "//compilers/llvm:opt_util",
         "//deeplearning/deeptune/opencl/heterogeneous_mapping:utils",
-        "//deeplearning/ml4pl/graphs/unlabelled/llvm2graph/cfg:llvm_util",
+        "//deeplearning/ml4pl/graphs/llvm2graph/legacy/cfg:llvm_util",
         "//deeplearning/ncc:task_utils",
         "//deeplearning/ncc:vocabulary",
         "//labm8/py:app",
@@ -102,7 +102,7 @@ py_test(
         ":models",
         ":testlib",
         "//deeplearning/deeptune/opencl/heterogeneous_mapping:conftest",
-        "//deeplearning/ml4pl/graphs/unlabelled/llvm2graph/cfg:llvm_util",
+        "//deeplearning/ml4pl/graphs/llvm2graph/legacy/cfg:llvm_util",
         "//labm8/py:test",
         "//third_party/py/networkx",
         "//third_party/py/numpy",
